undetections; final outcome tests
adds method for handling "undetections", so now infections can lead to more that one detection event, and these are counted separately. E.g., an infection fails out of treatment, then is screened again and successfully treated with Y. in the final tally this means detections + 2, fails +1 and success with Y + 1.

also within infection processing, a hashmap checks that only one of the mutually-exclusive outcomes is recorded for each infection. (failures + undetections are processed separately, as described above.)